stdcommands/ce: use new yaml library
Use the newly introduced github.com/goccy/go-yaml library package to
parse YAML and JSON in the customembed command.

I had to change parsed to no longer be a pointer type, because there
were some issues with Unmarshalling into said struct.

As an added bonus, we can now throw either YAML and JSON at
yaml.Unmarshal, because a) JSON is a subset of YAML and b) it just
works.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>