moderation: move SaveConfig, GetConfigOrDefault to config.go
Previously, these functions (which directly accessed database) were in plugin_bot.go. However, they are safe to use in web code as well (in fact, they are the *only* config-related functions safe to use in web code.) To make this distinction clearer, this commit moves them to a more neutral file.