tickets: improve Missing Access error message
Improve the "Missing Access" error message when trying to send a
ticket's transcript and/or images to the ticket logs channel.

Currently, the bot simply responds with a quite generic "Missing Access"
error, which, whilst technically correct, isn't exactly helpful.
Therefore, send a proper error instead that explains what went wrong and
that we're refusing to close the ticket for exactly that reason.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>