interactions: multiple modal specific error
This additional specific error for modal responses differentiates it
from the standard multiple interaction responses error - which
encourages users to consider using a followup and would be an
inappropriate solution for a user attempting to use two modal responses.
This aims to fortify the validity of the suggestion to consider a
followup, since it makes the generic too many responses error only
present itself when what a user is attempting could reasonably be
accomplished by using followups instead.

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>