feat(youtube): Add error for missing path
A URL without a path is technically a valid URL,
so shouldn't throw an error when being parsed, but we know it can't
be a valid YouTube URL, so we can exit early if we have recieved
an empty path. It also allows us to make some assumptions about our data
later.