fix(rolecommands): Re-add `rr` check on menu setup
When the `-rr` flag was introduced, a condition to flip its state to
`true` if the `-rr` flag was included, was introduced.
For whatever reason, at some point this conditional was removed, and the
flag, has not been functional when creating a Role Menu ever since

This commit introduces an inverted version of the statement used for the
`-nodm`, for `-rr`, to hopefully reintroduce this functionality.