fix(youtube): Fix some path parsing quirks (#1767)
* feat(youtube): Add error for missing path

A URL without a path is technically a valid URL,
so shouldn't throw an error when being parsed, but we know it can't
be a valid YouTube URL, so we can exit early if we have recieved
an empty path. It also allows us to make some assumptions about our data
later.

* fix(youtube): Allow ULRs with 2+ segments

Previously I assumed parsed URLs wouldn't have more than two segments,
but this is incorrect, notably in the case of channels, where each
section of a channel adds a third segment to the URL.
This removes this restriction, as it shouldn't dramatically affect
performance anyway.

* fix(youtube): Split path before handle parsing

As noted with the prior commit, each section of a YouTube channel page,
adds a segment to the channel URL. As the current implementation doesn't
account for this, it will flag these paths as invalid, even if they
aren't. This commit moves the existing function to split the string,
before the handle parsing check, so the first segment can be checked
separately from the rest of the path.