feat(reddit): Trim `/r/` from new subreddit feeds
As the Reddit feed is designed around using  the subreddit name
exclusively, submitting data other than just the subreddit name,
is erroneous. Currently, as the control panel notes, this includes
submitting the subreddit name with a leading `/r/`.

As this error is common enough to require a note about it
in the control panel, and mitigating against it is trivial,
this commit updates handling of new feeds,
to strip all instances of these leading characters.