refact(reddit): Update note on subreddit input
As referenced in the prior commit, the control panel made a note
about the appropriate formatting when supplying subreddit names.
This commit makes a simple change, to update the note to clarify,
a subreddit name should be submitted, rather than a URL,
as opposed to warning about cases of leading `/r/`, as these
should now be stripped.