refact(reddit): Clean up subreddit name processing (#1790)
* feat(reddit): Trim `/r/` from new subreddit feeds

As the Reddit feed is designed around using  the subreddit name
exclusively, submitting data other than just the subreddit name,
is erroneous. Currently, as the control panel notes, this includes
submitting the subreddit name with a leading `/r/`.

As this error is common enough to require a note about it
in the control panel, and mitigating against it is trivial,
this commit updates handling of new feeds,
to strip all instances of these leading characters.

* refact(reddit): Update note on subreddit input

As referenced in the prior commit, the control panel made a note
about the appropriate formatting when supplying subreddit names.
This commit makes a simple change, to update the note to clarify,
a subreddit name should be submitted, rather than a URL,
as opposed to warning about cases of leading `/r/`, as these
should now be stripped.

* refact(reddit): Stop name processing on update

For whatever reason, the function for modifying Reddit feeds,
has been stripping the whitespace around subreddit names,
and converting them to lowercase again, whenever a subreddit feed
was modified, even though the function used when adding feeds
had already been doing so.

This commit removes these redundant function.

* refact(reddit): Stop name processing on remove

Similarly to the previous commit, this one also removes
redundant function calls used when removing subreddit feeds,
due to the relevant data being processed when adding those feeds.

* refact(reddit): Remove iterative `r/` prefix trim

When initially implementing `3aa0b17`, I was unsure of the validity
of iteratively trimming instances of `r/` from the start
of the subreddit name, and it was confirmed this is not preferrable.
So, this commit removes the `for` loop, in favour of a single operation.