tickets: Switch to variable for closing message
It has been highlighted by one uses, that the lack of granularity
from the closing message, concerns them.

Consequently, this commit updates the ticket closing function,
to use a variable for the closing message, as opposed to the const. msg.
used currently.
This acts largely as a prelude to the subsequent commits,
which will add more to the message, conditionally,
based on the logs YAGPDB will create.