tickets: Drop ticket closing response from buttons
Due to the design of the ticket button handling functions,
closing tickets resulted in the ticket closing prompt
being doubled up on with an ephemeral response.

For this reason, these prompts served little purpose,
as a prompt was already created as a normal message,
as part of the `tickets.closeTicket` function.
Consquently, this effectively just introduced a dependency
on the `tickets.closingTicketMsg` constant, that this series of commits,
served to remove.