dgo/dstate: add RoleSubscriptionData to Message struct (#1865)
* discordgo: add RoleSubscriptionData message field

Discord populates a RoleSubscriptionData object on messages of type 25
(ROLE_SUBSCRIPTION_PURCHASE); add it to discordgo.Message such that we
can make it available in custom commands (and elsewhere).

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* dstate: add RoleSubscriptionData to MessageState struct

The previous commit introduced a new RoleSubscriptionData field to
discordgo.Message -- also add it to MessageState, such that it can be
used in custom commands et al.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>