templates: properly edit message flags in tmplEditMessage (#1885)
The tmplEditMessage function does not properly transfer message flags
over to the new message object, causing some funny behaviour when, for
example, trying to set `suppress_embeds` to `true` via the
`complexMessageEdit`-builder. In other words, it does not work at all.

Copying the message flags over to the new messag fixes this.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>