Fix typo in the error message on line 84 (` replaced with ') (#1891)
The error message on line 84 has a backtick (`) instead of a single quote ('), which is inconsistent with the rest of the error message. Although this has no impact on functionality, this can cause formatting issues. If users use backticks anywhere in their response for parseArgs, the response may look messy.