evalcc: raise ops limit
Raise the ops limit on evalcc to 200k and 500k respectively; it can be
argued that, because evalcc is about a fifth of a custom command, the
ops limit should be a fifth of a custom command, as well.

For reference, the generic api call counter is already at 20, which is a
fifth of the normal custom command api call limit.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>