logs: allow selection of channel allow/denylist mode
Allow the user to make a choice between an allow- or denylist mode for
the channel list on the logs configuration page, similar to how custom
commands can be configured.

The old behaviour of it being an allowlist is retained, so users will
have to opt-in to the allowlist behaviour at their discretion.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>