Undo making sendPunishDM execute as a go routine when kicking or banning (#1926)
Making the sendPunishDM execute as a go routine when kicking or banning generates a potential problem. If a member is in the server, then is kicked/banned, they will not receive the punishment DM if they are removed from the server before the punishment template has finished executing. Not having it run as a go routine will help fix this issue.